attempt2

+++ internal/winapi/jobobject.go
@@ -163,7 +163,7 @@ type JOBOBJECT_ASSOCIATE_COMPLETION_PORT struct {
 
 // https://learn.microsoft.com/en-us/windows/win32/api/winnt/ns-winnt-group_affinity
 type JOBOBJECT_CPU_GROUP_AFFINITY struct {
-	CpuMask  uintptr
+	CpuMask  uint64
 	CpuGroup uint16
 	Reserved [3]uint16
 }
